Refactoring the translate process
