+++ onmt/__init__.py
@@ -1,13 +1,11 @@
 import onmt.io
-import onmt.translate
 import onmt.Models
 import onmt.Loss
+import onmt.translate
 import onmt.opts
-from onmt.ServerModels import ServerModels, ServerModelError
 from onmt.Trainer import Trainer, Statistics
 from onmt.Optim import Optim
 
 # For flake8 compatibility
 __all__ = [onmt.Loss, onmt.Models, onmt.opts,
-           Trainer, Optim, Statistics, onmt.io, onmt.translate,
-           ServerModels, ServerModelError]
+           Trainer, Optim, Statistics, onmt.io, onmt.translate]
