+++ onmt/translate/__init__.py
@@ -2,7 +2,8 @@ from onmt.translate.Translator import Translator
 from onmt.translate.Translation import Translation, TranslationBuilder
 from onmt.translate.Beam import Beam, GNMTGlobalScorer
 from onmt.translate.Penalties import PenaltyBuilder
+from onmt.translate.TranslationServer import TranslationServer
 
 __all__ = [Translator, Translation, Beam,
            GNMTGlobalScorer, TranslationBuilder,
-           PenaltyBuilder]
+           PenaltyBuilder, TranslationServer]
